pyproject.toml
[project]
name = "cs-space"
version = "0.1.0"
description = ""
authors = [
    {name = "JonasBuffington"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all] (>=0.120.3,<0.121.0)",
    "neo4j (>=6.0.2,<7.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "typer (>=0.20.0,<0.21.0)",
    "rich (>=14.2.0,<15.0.0)",
    "google-genai (>=1.47.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "python-dotenv (>=1.2.1,<2.0.0)"
]
